requirements.txt
# Web framework
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
python-multipart==0.0.9
starlette==0.36.3

# Google AI
google-genai==1.60.0
google-generativeai==0.8.6
google-cloud-texttospeech==2.16.3

# Database (optional — chatbot works without it)
SQLAlchemy==2.0.29
psycopg2-binary==2.9.11

# PDF generation
reportlab==4.2.2

# Utilities
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1

# Type checking / validation
pydantic==2.12.5
annotated-types==0.7.0
